Fix #59: Input#text => textarea
This change converts the message input to a textarea for multiline
support. It also ups the font-size to 16px on the textarea so that
zooming on focus on mobile safari is disabled, which breaks the
responsive layout (and makes it easier to read). Add a button that
triggers the send operation. Persist the message on local state in the
form component, which we'll use when we add image attachment support.