Fix #77: Responsive Header
This change creates a responsive header for the app that lives at the
top of the site on mobile and as a sidebar for larger desktop screens.
It converts the primary elements in the layout to use the new flexbox
api. It also sets a fixed height on the messages container to allow for
scrolling when the message list overflows the height of the container,
which is less than flexy but will work for now.