Added initial pincode logic
When user invited to a room whose URL hash ends with '--', ask for
pincode. Otherwise, don't ask for pincode.

Still need to somehow tell users/room creators about this convention
so they know they can use it simply by appending '--' to a LeapChat
URL.

Pincode (which can be any string) currently not stored anywhere, even
in this.state, for maximum security.  Could later store in
localStorage; originally just added it to the end of
`document.location.hash`, but that's a bad idea for multiple reasons.